requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.59.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
h5py==3.10.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.6.2
MedPy==0.5.2
mpmath==1.3.0
networkx==3.3
numpy==1.23.5
opencv-python==4.9.0.80
packaging==25.0
pathtools==0.1.2
Pillow==9.3.0
promise==2.3
protobuf==4.25.8
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
scipy==1.15.3
sentry-sdk==2.34.0
setproctitle==1.3.6
shortuuid==1.0.13
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tqdm==4.64.1
typing_extensions==4.14.1
urllib3==2.5.0
wandb==0.13.5
